S3/Cloudfront: fix bug caused by AWS API change (#89)
Amazon changed their API in a way that caused a change in behavior of
older API versions, including the one used by s3cmd (2010-11-01.) As
noted in several bug reports on other projects[1][2] the S3Origin
object is required to have an OriginAccessIdentity field.

This commit is a quick hack that adds an empty OriginAccessIdentity
field to the default config. It does not allow it to be set.

[1] https://bitbucket.org/jmurty/jets3t/issue/150/update-to-api-version-2012-05-05
[2] https://github.com/aws/aws-sdk-ruby/issues/144